Seed with tflow seq2seq
Seeding development with a working seq2seq (with attention) model from
the Tensor Flow library built to learn translations between english and
french.

Also built a simple simulator of trace data. These traces need to be
reformatted and/or the seq2seq preprocessor needs to be modified to
read them.