Be able to decode traces with model
Modified the existing decode function to take traces in their current
form and produce decoded output. So it can now actually decode using
the model. Terrible, terrible base-calls but excitingly the whole thing
can be trained on trace input and based on that model produce DNA
sequence output. I’ll let it train on the cloud for a bit and see how
it performs. But there is obviously going to be more subtlety to it
than this.

The next task is, given a file of decodings and a file of sequences,
compute the length-normed levenshtein distance averaged across all of
the sequences.